Rename rustdoc js test suites
