--- src/test/rustdoc-js-not-std/basic.js
@@ -1,7 +1,0 @@
-const QUERY = 'Fo';
-
-const EXPECTED = {
-    'others': [
-        { 'path': 'basic', 'name': 'Foo' },
-    ],
-};
