+++ src/test/rustdoc-js-std/alias-1.js
