+++ src/test/rustdoc-js-std/alias-2.js
