+++ src/test/rustdoc-js-std/alias-3.js
