+++ src/test/rustdoc-js-std/alias.js
