+++ src/test/rustdoc-js-std/fn-forget.js
