+++ src/test/rustdoc-js-std/from_u.js
