+++ src/test/rustdoc-js-std/macro-check.js
