+++ src/test/rustdoc-js-std/macro-print.js
