+++ src/test/rustdoc-js-std/multi-query.js
