+++ src/test/rustdoc-js-std/never.js
