+++ src/test/rustdoc-js-std/should-fail.js
