+++ src/test/rustdoc-js-std/string-from_ut.js
