+++ src/test/rustdoc-js-std/struct-vec.js
