+++ src/test/rustdoc-js-std/vec-new.js
