+++ src/test/rustdoc-js/basic.js
@@ -1,15 +1,7 @@
-const QUERY = 'String';
+const QUERY = 'Fo';
 
 const EXPECTED = {
     'others': [
-        { 'path': 'std::string', 'name': 'String' },
-        { 'path': 'std::ffi', 'name': 'CString' },
-        { 'path': 'std::ffi', 'name': 'OsString' },
-    ],
-    'in_args': [
-        { 'path': 'std::str', 'name': 'eq' },
-    ],
-    'returned': [
-        { 'path': 'std::string::String', 'name': 'add' },
+        { 'path': 'basic', 'name': 'Foo' },
     ],
 };
